refactor: reorganize ruby-generator tests for v2 as default
Extract v1-specific tests to v1/ subdirectory:
- v1/class.test.js: stage v1 and withSpriteNew v1 tests
- v1/file-output-patterns.test.js: v1 no-class and @ruby:class sections
- v1/fuzz-hunt.test.js: v1 extension output and self.when hat block tests

Update main test files to use v2 syntax:
- self.when(:flag_clicked) → when_flag_clicked
- targetToCode() → targetToCode(target, {version: '2'})
- def self.add → def add
- Renumber section comments after v1 extraction
